Convert timestamp from uint256_t to uint64_t

+++ src/libDirectoryService/DSBlockPreProcessing.cpp
@@ -728,13 +728,13 @@ bool DirectoryService::RunConsensusOnDSBlockWhenDSPrimary() {
   // TODO: Revise DS block structure
   {
     lock_guard<mutex> g(m_mediator.m_mutexCurSWInfo);
-    m_pendingDSBlock.reset(
-        new DSBlock(DSBlockHeader(dsDifficulty, difficulty, prevHash,
-                                  m_mediator.m_selfKey.second, blockNum,
-                                  m_mediator.m_currentEpochNum,
-                                  get_time_as_int(), m_mediator.m_curSWInfo,
-                                  powDSWinners, dsBlockHashSet, committeeHash),
-                    CoSignatures(m_mediator.m_DSCommittee->size())));
+    m_pendingDSBlock.reset(new DSBlock(
+        DSBlockHeader(
+            dsDifficulty, difficulty, prevHash, m_mediator.m_selfKey.second,
+            blockNum, m_mediator.m_currentEpochNum,
+            static_cast<uint64_t>(get_time_as_int()), m_mediator.m_curSWInfo,
+            powDSWinners, dsBlockHashSet, committeeHash),
+        CoSignatures(m_mediator.m_DSCommittee->size())));
     m_pendingDSBlock->SetBlockHash(m_pendingDSBlock->GetHeader().GetMyHash());
   }
